+++ changelog.md
@@ -4,6 +4,7 @@
 
 - jcontent_by_categories : caché l'article en cours et article en cours indéterminé : essayer de le récupérer par getInput
 - sql : ajout pagination, setlimit accepte un nombre ou une requête sql
+- sql : ajout de route-format : calcul du lien vers un article
 
 ## 08/11/2025 - version 5.4.4